+++ wrapper/src/fingerprint.cc
@@ -24,11 +24,11 @@ namespace RDKit {
         return bitvect->getNumOnBits();
     }
 
-    std::unique_ptr<std::vector<uint8_t>> explicit_bit_vect_to_bytes_vec(std::shared_ptr<ExplicitBitVect> bitvect) {
-        std::vector<uint8_t> bytes;
+    std::unique_ptr<std::vector<uint64_t>> explicit_bit_vect_to_bytes_vec(std::shared_ptr<ExplicitBitVect> bitvect) {
+        std::vector<uint64_t> bytes;
         bytes.reserve(bitvect->dp_bits->num_blocks());
         boost::to_block_range(*bitvect->dp_bits, (std::back_inserter(bytes)));
-        std::vector<uint8_t> *bytes_heap = new std::vector<uint8_t>(bytes);
-        return std::unique_ptr<std::vector<uint8_t>>(bytes_heap);
+        std::vector<uint64_t> *bytes_heap = new std::vector<uint64_t>(bytes);
+        return std::unique_ptr<std::vector<uint64_t>>(bytes_heap);
     }
 }
